+++ src/checks/incorrect_delimiter.rs
@@ -48,50 +48,26 @@ impl Check for IncorrectDelimiterChecker<'_> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use std::path::PathBuf;
+    use crate::common::tests::*;
 
     #[test]
     fn working_run() {
         let mut checker = IncorrectDelimiterChecker::default();
-        let line = LineEntry {
-            number: 1,
-            file: FileEntry {
-                path: PathBuf::from(".env"),
-                file_name: ".env".to_string(),
-                total_lines: 1,
-            },
-            raw_string: String::from("FOO_BAR=FOOBAR"),
-        };
+        let line = line_entry(1, 1, "FOO_BAR=FOOBAR");
         assert_eq!(None, checker.run(&line));
     }
 
     #[test]
     fn working_with_digits_run() {
         let mut checker = IncorrectDelimiterChecker::default();
-        let line = LineEntry {
-            number: 1,
-            file: FileEntry {
-                path: PathBuf::from(".env"),
-                file_name: ".env".to_string(),
-                total_lines: 1,
-            },
-            raw_string: String::from("F1OO=BAR"),
-        };
+        let line = line_entry(1, 1, "F1OO=BAR");
         assert_eq!(None, checker.run(&line));
     }
 
     #[test]
     fn incorrect_leading_char() {
         let mut checker = IncorrectDelimiterChecker::default();
-        let line = LineEntry {
-            number: 1,
-            file: FileEntry {
-                path: PathBuf::from(".env"),
-                file_name: ".env".to_string(),
-                total_lines: 1,
-            },
-            raw_string: String::from("*FOO=BAR"),
-        };
+        let line = line_entry(1, 1, "*FOO=BAR");
         // expect None because this warning should be found by LeadingCharacterChecker
         assert_eq!(None, checker.run(&line));
     }
@@ -99,15 +75,7 @@ mod tests {
     #[test]
     fn incorrect_leading_chars_and_invalid_delimiter() {
         let mut checker = IncorrectDelimiterChecker::default();
-        let line = LineEntry {
-            number: 1,
-            file: FileEntry {
-                path: PathBuf::from(".env"),
-                file_name: ".env".to_string(),
-                total_lines: 1,
-            },
-            raw_string: String::from("***F-OOBAR=BAZ"),
-        };
+        let line = line_entry(1, 1, "***F-OOBAR=BAZ");
 
         let expected = Some(Warning::new(
             line.clone(),
@@ -121,15 +89,7 @@ mod tests {
     #[test]
     fn incorrect_ending_delimiter() {
         let mut checker = IncorrectDelimiterChecker::default();
-        let line = LineEntry {
-            number: 1,
-            file: FileEntry {
-                path: PathBuf::from(".env"),
-                file_name: ".env".to_string(),
-                total_lines: 1,
-            },
-            raw_string: String::from("FOO*=BAR"),
-        };
+        let line = line_entry(1, 1, "FOO*=BAR");
 
         let expected = Some(Warning::new(
             line.clone(),
@@ -143,15 +103,7 @@ mod tests {
     #[test]
     fn failing_run() {
         let mut checker = IncorrectDelimiterChecker::default();
-        let line = LineEntry {
-            number: 1,
-            file: FileEntry {
-                path: PathBuf::from(".env"),
-                file_name: ".env".to_string(),
-                total_lines: 1,
-            },
-            raw_string: String::from("FOO-BAR=FOOBAR"),
-        };
+        let line = line_entry(1, 1, "FOO-BAR=FOOBAR");
         let expected = Some(Warning::new(
             line.clone(),
             "IncorrectDelimiter",
@@ -163,15 +115,7 @@ mod tests {
     #[test]
     fn failing_with_whitespace_run() {
         let mut checker = IncorrectDelimiterChecker::default();
-        let line = LineEntry {
-            number: 1,
-            file: FileEntry {
-                path: PathBuf::from(".env"),
-                file_name: ".env".to_string(),
-                total_lines: 1,
-            },
-            raw_string: String::from("FOO BAR=FOOBAR"),
-        };
+        let line = line_entry(1, 1, "FOO BAR=FOOBAR");
         let expected = Some(Warning::new(
             line.clone(),
             "IncorrectDelimiter",
@@ -183,15 +127,7 @@ mod tests {
     #[test]
     fn unformatted_run() {
         let mut checker = IncorrectDelimiterChecker::default();
-        let line = LineEntry {
-            number: 1,
-            file: FileEntry {
-                path: PathBuf::from(".env"),
-                file_name: ".env".to_string(),
-                total_lines: 1,
-            },
-            raw_string: String::from("FOO-BAR"),
-        };
+        let line = line_entry(1, 1, "FOO-BAR");
         // there's no key, so KeyWithoutValueChecker should catch this error
         assert_eq!(None, checker.run(&line));
     }
@@ -199,15 +135,7 @@ mod tests {
     #[test]
     fn trailing_space_run() {
         let mut checker = IncorrectDelimiterChecker::default();
-        let line = LineEntry {
-            number: 1,
-            file: FileEntry {
-                path: PathBuf::from(".env"),
-                file_name: ".env".to_string(),
-                total_lines: 1,
-            },
-            raw_string: String::from("FOO_BAR =FOOBAR"),
-        };
+        let line = line_entry(1, 1, "FOO_BAR =FOOBAR");
         // has a trailing space, so SpaceCharacterChecker should catch this error
         assert_eq!(None, checker.run(&line));
     }
@@ -215,30 +143,14 @@ mod tests {
     #[test]
     fn empty_run() {
         let mut checker = IncorrectDelimiterChecker::default();
-        let line = LineEntry {
-            number: 1,
-            file: FileEntry {
-                path: PathBuf::from(".env"),
-                file_name: ".env".to_string(),
-                total_lines: 1,
-            },
-            raw_string: String::from(""),
-        };
+        let line = line_entry(1, 1, "");
         assert_eq!(None, checker.run(&line));
     }
 
     #[test]
     fn short_run() {
         let mut checker = IncorrectDelimiterChecker::default();
-        let line = LineEntry {
-            number: 1,
-            file: FileEntry {
-                path: PathBuf::from(".env"),
-                file_name: ".env".to_string(),
-                total_lines: 1,
-            },
-            raw_string: String::from("F=BAR"),
-        };
+        let line = line_entry(1, 1, "F=BAR");
         assert_eq!(None, checker.run(&line));
     }
 }